chore: Add validation for route coordinates.
The backend will check on POST and UPDATE any given coordinates and
return a BAD_REQUEST if there are less than two coordinates,
more than 1.000.000 and/or if the dimensions are mixed (2d, 3d coordinates
in the same route).
The checks are added to avoid hitting an error while inserting into the
database.