chore!: rework backend code to be more modular and have proper architecture
while the code before looked okay it suffered from multiple issues. The biggest one was that service and persistance layers were mingled and there was no clear distinction between DTO and entites. Furthermore the code is made more heterogene by unifying naming of files and the overall structure of files