chore!: update backend architecture to use a seperate database layer (#21)
* chore!: rework backend code to be more modular and have proper architecture

while the code before looked okay it suffered from multiple issues. The biggest one was that service and persistance layers were mingled and there was no clear distinction between DTO and entites. Furthermore the code is made more heterogene by unifying naming of files and the overall structure of files

* chore: add missing images files

* chore: fixing tests and linter issues

* chore: clean folder structure

* chore: add more ci cd tools and refactor tests

Signed-off-by: Gil Engel <gil.engel@web.de>

---------

Signed-off-by: Gil Engel <gil.engel@web.de>